[#24] Refactoring
