+++ app/src/androidTest/java/com/justsoft/redditshareinterceptor/utils/VolleyRequestHelperTest.kt
@@ -1,8 +1,8 @@
-package com.justsoft.redditshareinterceptor.util
+package com.justsoft.redditshareinterceptor.utils
 
 import androidx.test.platform.app.InstrumentationRegistry
 import com.android.volley.toolbox.Volley
-import com.justsoft.redditshareinterceptor.util.request.VolleyRequestHelper
+import com.justsoft.redditshareinterceptor.utils.request.VolleyRequestHelper
 import org.junit.Assert.assertEquals
 import org.junit.Test
