+++ app/src/main/java/com/justsoft/redditshareinterceptor/UniversalMediaDownloader.kt
@@ -1,11 +1,11 @@
 package com.justsoft.redditshareinterceptor
 
 import android.net.Uri
-import com.justsoft.redditshareinterceptor.downloaders.MultipleFileDownloader
-import com.justsoft.redditshareinterceptor.downloaders.SingleFileDownloader
 import com.justsoft.redditshareinterceptor.model.ProcessingProgress
 import com.justsoft.redditshareinterceptor.model.media.MediaDownloadInfo
-import com.justsoft.redditshareinterceptor.util.request.RequestHelper
+import com.justsoft.redditshareinterceptor.utils.downloaders.MultipleFileDownloader
+import com.justsoft.redditshareinterceptor.utils.downloaders.SingleFileDownloader
+import com.justsoft.redditshareinterceptor.utils.request.RequestHelper
 import java.io.OutputStream
 
 class UniversalMediaDownloader(
