+++ app/src/main/java/com/justsoft/redditshareinterceptor/processors/PostProcessor.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import com.justsoft.redditshareinterceptor.model.RedditPost
 import com.justsoft.redditshareinterceptor.model.media.MediaContentType
 import com.justsoft.redditshareinterceptor.model.media.MediaDownloadObject
-import com.justsoft.redditshareinterceptor.util.request.RequestHelper
+import com.justsoft.redditshareinterceptor.utils.request.RequestHelper
 
 /**
  * An interface for processing different variations of posts
