+++ app/src/main/java/com/justsoft/redditshareinterceptor/processors/RedditImagePostProcessor.kt
@@ -4,8 +4,8 @@ import android.os.Bundle
 import com.justsoft.redditshareinterceptor.model.RedditPost
 import com.justsoft.redditshareinterceptor.model.media.MediaContentType
 import com.justsoft.redditshareinterceptor.model.media.MediaDownloadObject
-import com.justsoft.redditshareinterceptor.util.request.RequestHelper
-import com.justsoft.redditshareinterceptor.util.urlDecode
+import com.justsoft.redditshareinterceptor.utils.request.RequestHelper
+import com.justsoft.redditshareinterceptor.utils.urlDecode
 import org.json.JSONObject
 
 class RedditImagePostProcessor : PostProcessor {
