+++ app/src/main/java/com/justsoft/redditshareinterceptor/utils/Stopwatch.kt
@@ -1,4 +1,4 @@
-package com.justsoft.redditshareinterceptor.util
+package com.justsoft.redditshareinterceptor.utils
 
 class Stopwatch {
