+++ app/src/main/java/com/justsoft/redditshareinterceptor/utils/downloaders/SingleFileDownloader.kt
@@ -1,11 +1,11 @@
-package com.justsoft.redditshareinterceptor.downloaders
+package com.justsoft.redditshareinterceptor.utils.downloaders
 
 import android.net.Uri
 import android.util.Log
 import com.justsoft.redditshareinterceptor.model.ProcessingProgress
 import com.justsoft.redditshareinterceptor.model.media.MediaDownloadObject
 import com.justsoft.redditshareinterceptor.model.media.metadata.MediaMetadata
-import com.justsoft.redditshareinterceptor.util.request.RequestHelper
+import com.justsoft.redditshareinterceptor.utils.request.RequestHelper
 import java.io.OutputStream
 import java.net.HttpURLConnection
 import java.net.URL
